% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validate-glmm-lasso.R
\name{cross_validate_glmm_lasso}
\alias{cross_validate_glmm_lasso}
\alias{is_cv_glmm_lasso}
\alias{summary.cv_glmm_lasso}
\title{Preform Cross-Validated \code{glmm} Lasso}
\usage{
cross_validate_glmm_lasso(
  data,
  fixed,
  random,
  family,
  folds = 10L,
  .repeats = 1,
  .lambda,
  decision_threshold = 0.5,
  subsample = NULL,
  r_seed = 1234,
  save_fits = FALSE,
  save_splits = FALSE,
  ...
)

is_cv_glmm_lasso(x)

\method{summary}{cv_glmm_lasso}(object, ci_alpha = 0.05, ...)
}
\arguments{
\item{data}{The \code{data.frame} containing analysis data as columns.}

\item{fixed}{The formula for the fixed effect. For
classification problems, the LHS of the fixed formula (endpoint) needs to be
re-coded to 1 or 0 (not strings or factors).}

\item{random}{List. Indicating the full formula for the random
effect. For \code{glmmLasso} the formula  should be of the format
\code{list(randomEffectTerm = ~1)}.}

\item{family}{\code{character(1)}. Which family of models to fit to the
generalized linear mixed model? This should be a \code{glm} family.
If no family is provided, a linear mixed
model is fit instead of a generalized linear mixed model.}

\item{folds}{\code{integer(1)}. Number of folds to perform (k-fold cross-validation).}

\item{.repeats}{\code{integer(1)}. Number of repeats to perform as part of
repeated k-fold cross-validation. Default is 1 (no repeats).}

\item{.lambda}{The values for lambda in the lasso:
default is \verb{[0, 0.01, 0.1, 1, 10, 100, 1000]}.}

\item{decision_threshold}{\code{numeric(1)}. The decision threshold
to use for classification.}

\item{subsample}{\code{character(1)}. The type of sub-sampling to perform.
See \code{\link[splyr:rebalance]{splyr::rebalance()}} \verb{method =} argument.
Defaults to "none" (\code{NULL}).}

\item{r_seed}{\code{integer(1)}. Value to set the random seed.}

\item{save_fits}{\code{logical(1)}. Should individual model fits for each
\verb{<repeat/fold/tuning parameter>} iteration be saved?}

\item{save_splits}{\code{logical(1)}. Should individual split
data sets for each \verb{<repeat/fold/tuning parameter>}
iteration be saved? Is the output of \code{\link[rsample:vfold_cv]{rsample::vfold_cv()}}
such that the splits need to be passed to
\code{\link[rsample:as.data.frame.rsplit]{rsample::analysis()}} or \code{\link[rsample:as.data.frame.rsplit]{rsample::assessment()}} to be accessed?}

\item{...}{Additional arguments passed to the underlying
\code{\link[glmmLasso:glmmLasso]{glmmLasso::glmmLasso()}} model fitting function.}

\item{x}{An object to be tested for class \code{cv_glmm_lasso}.}

\item{object}{A \code{cv_glmm_lasso} class object.}

\item{ci_alpha}{\code{double(1)}. Significance level of the confidence
interval for \code{alpha}.}
}
\value{
A \code{cv_glmm_lasso} class object with components:
\item{metrics}{A \code{tibble} of cross-validation metrics.}
\item{formula}{The mixed-effects model formula.}
\item{model_type}{Either \code{classification} or \code{regression}.}
\item{random_seed}{Random seed for the cross-validation folds.}

The \code{metrics} entry is a \code{tibble} where the number of rows is equal
to the cross of \code{.repeats}, \code{folds}, and \code{.lambda} and
contains the following columns:
\itemize{
\item \code{repeats} (id)
\item \code{folds}   (id2)
\item \code{.lambda}
\item \code{Sensitivity}
\item \code{Specificity}
\item \code{auc}
\item \code{.split} (optional)
\itemize{
\item A list column of \code{tibbles} of the data splits: \code{"vfold_cv"}, \code{"rsplit"}
}
\item \code{fit} (optional)
\itemize{
\item A list column of model fit objects (\code{glmmLasso})
}
}
}
\description{
Fit cross-validated models for \code{\link[glmmLasso:glmmLasso]{glmmLasso::glmmLasso()}}.

For \code{is_cv_glmm_lasso}: A logical test for
the \code{cv_glmm_lasso} class.
}
\section{Functions}{
\itemize{
\item \code{summary(cv_glmm_lasso)}: S3 summary method for class \code{cv_glmm_lasso}.

}}
\seealso{
\code{\link[=fit_glmm_lasso]{fit_glmm_lasso()}}

Other GLMMlasso: 
\code{\link{fit_glmm_lasso}()}
}
\author{
Gargi Datta & Stu Field
}
\concept{GLMMlasso}
