% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-glmm-lasso.R
\name{fit_glmm_lasso}
\alias{fit_glmm_lasso}
\title{Fit a \code{glmmLasso} object from a \code{cv_glmm_lasso} object}
\usage{
fit_glmm_lasso(
  object,
  data,
  metric = c("auc", "sensitivity", "specificity", "rsqTrad", "rsqAdj", "rsqPred"),
  verbose = interactive(),
  ...
)
}
\arguments{
\item{object}{A \code{cv_glmm_lasso} class object to use to fit
a \code{glmm_lasso} object.}

\item{data}{Data to fit model with. Typically a \code{data.frame} object.}

\item{metric}{Which performance metric to use to choose
lambda for lasso models.}

\item{verbose}{\code{logical(1)}. Toggle verbosity.}

\item{...}{Additional parameters passed to \code{\link[glmmLasso:glmmLasso]{glmmLasso::glmmLasso()}}
model fitting function.}
}
\value{
A \code{glmmLasso} class object. See \code{\link[glmmLasso:glmmLasso]{glmmLasso::glmmLasso()}}.
}
\description{
From a \code{cv_glmm_lasso} object, find the best parameters from the cross-
validation and re-fit via \code{\link[glmmLasso:glmmLasso]{glmmLasso::glmmLasso()}}.
See \verb{[cross_validate_glmm_lasso()]}.
}
\seealso{
Other GLMMlasso: 
\code{\link{cross_validate_glmm_lasso}()}
}
\concept{GLMMlasso}
