% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGlmmLasso.R
\name{fitGlmmLasso}
\alias{fitGlmmLasso}
\title{Fit a \code{glmmLasso} object from a \code{cvGlmmLasso} object}
\usage{
fitGlmmLasso(
  object,
  data,
  metric = c("auc", "sensitivity", "specificity", "rsqTrad", "rsqAdj", "rsqPred"),
  verbose = interactive(),
  ...
)
}
\arguments{
\item{object}{A \code{cvGlmmLasso} class object to use to fit a \code{glmmLasso} object.}

\item{data}{Data to fit model with. Typically a \code{soma_adat} class object.}

\item{metric}{Which performance metric to use to choose lambda for lasso models.}

\item{verbose}{Logical. Toggle verbosity.}

\item{...}{Additional parameters passed to \code{\link[glmmLasso:glmmLasso]{glmmLasso::glmmLasso()}}
model fitting function.}
}
\value{
A \code{glmmLasso} class object.
}
\description{
From a \code{cvGlmmLasso} object, find the best parameters from the cross-
validation and re-fit via \code{\link[glmmLasso:glmmLasso]{glmmLasso::glmmLasso()}}. See \verb{[crossValidateGlmmLasso()]}.
}
\seealso{
Other GLMMlasso: 
\code{\link{crossValidateGlmmLasso}()}
}
\concept{GLMMlasso}
