% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_google.R
\name{geocode_google}
\alias{geocode_google}
\title{Geocodes one location through the Google API}
\usage{
geocode_google(stringlocation, info = c("lat", "lng"))
}
\arguments{
\item{stringlocation}{a location string}

\item{info}{the variables to inform: by default, c("lat","lng").
Information can be e.g. "locality","country","country_code","formatted_address".
It can also be any type supported by the Google Map geocoding API:
see \href{https://developers.google.com/maps/documentation/geocoding/intro#Types}{Address types}.}
}
\value{
tibble with additional columns corresponding to the info required
}
\description{
Geocodes one location through the Google API
}
\examples{
geocode_google("Grande Rue de la Guillotière, Lyon",
 info=c("lat",
        "lng",
        "country_code",
        "locality",
        "administrative_area_level_1"))
}
