% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_opencage.R
\name{geocode_opencage}
\alias{geocode_opencage}
\title{Geocodes one location through the OSM nominatim API}
\usage{
geocode_opencage(stringlocation, info = c("lat", "lng"))
}
\arguments{
\item{stringlocation}{a location string}

\item{info}{the variables to inform: by default, latlng.
Additional information can be e.g.
\itemize{
\item bounds.northeast.lat
\item bounds.northeast.lng
\item bounds.southwest.lat
\item bounds.southwest.lng
\item components.ISO_3166-1_alpha-2
\item components.ISO_3166-1_alpha-3
\item components._category
\item components._type
\item components.boundary
\item components.continent
\item components.country
\item components.country_code
\item components.state
\item confidence
\item formatted
\item annotations.FIPS.county
\item annotations.FIPS.state
\item annotations.currency.disambiguate_symbol
\item components.county
\item components.state_code
\item components.town
\item query
\item annotations.DMS.lat
\item annotations.DMS.lng
\item annotations.MGRS
\item annotations.Maidenhead
\item annotations.Mercator.x
\item annotations.Mercator.y
\item annotations.OSM.edit_url
\item annotations.OSM.note_url
\item annotations.OSM.url
\item annotations.UN_M49.statistical_groupings
\item annotations.callingcode
\item annotations.currency.alternate_symbols
\item annotations.currency.decimal_mark
\item annotations.currency.html_entity
\item annotations.currency.iso_code
\item annotations.currency.iso_numeric
\item annotations.currency.name
\item annotations.currency.smallest_denomination
\item annotations.currency.subunit
\item annotations.currency.subunit_to_unit
\item annotations.currency.symbol
\item annotations.currency.symbol_first
\item annotations.currency.thousands_separator
\item annotations.flag
\item annotations.geohash
\item annotations.qibla
\item annotations.roadinfo.drive_on
\item annotations.roadinfo.speed_in
\item annotations.sun.rise.apparent
\item annotations.sun.rise.astronomical
\item annotations.sun.rise.civil
\item annotations.sun.rise.nautical
\item annotations.sun.set.apparent
\item annotations.sun.set.astronomical
\item annotations.sun.set.civil
\item annotations.sun.set.nautical
\item annotations.timezone.name
\item annotations.timezone.now_in_dst
\item annotations.timezone.offset_sec
\item annotations.timezone.offset_string
\item annotations.timezone.short_name
\item annotations.what3words.words
\item annotations.wikidata
}}
}
\description{
Geocodes one location through the OSM nominatim API
}
\examples{
geocode_opencage("la Guillotière, Lyon",info=c("lat","lng","components.country_code"))
}
