% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expressions.R
\name{get_expressions}
\alias{get_expressions}
\title{Returns the table of expressions in a given language.
This function uses the IRaMuTeQ package's dictionnaries.}
\usage{
get_expressions(language = "fr")
}
\arguments{
\item{language.}{Can be one of "de" (German),"en" (English),"fr" (French),"gl" (Galician),"gr" (Greek),"it" (Italian),"pt" (Portuguese),"sp" (Spanish),"sw" (Swedish)}
}
\value{
tibble with word= word or expression, word_cor= corrected word or expression. For instance, in French, "aujourd'hui" is corrected as "ajourd_hui", and "curriculum vitae" is corrected as "curriculum_vitae".
}
\description{
Returns the table of expressions in a given language.
This function uses the IRaMuTeQ package's dictionnaries.
}
\examples{
fr_expr=get_expressions("fr")
}
