% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lexicon.R
\name{get_lexicon}
\alias{get_lexicon}
\title{Returns a lexicon in a given language.
This function uses the IRaMuTeQ package's dictionnaries.}
\usage{
get_lexicon(language = "fr")
}
\arguments{
\item{language}{Can be one of "
\itemize{
\item "de" (German),
\item "en" (English),
\item "fr" (French),
\item "gl" (Galician),
\item "gr" (Greek),
\item "it" (Italian),
\item "pt" (Portuguese),
\item "sp" (Spanish),
\item "sw" (Swedish)
}}
}
\value{
tibble with word, lemma and type.
}
\description{
Returns a lexicon in a given language.
This function uses the IRaMuTeQ package's dictionnaries.
}
\details{
Type: a word (in French lexicon) can be of type:
\itemize{
\item adj: adjective
\itemize{
\item adj_dem: demonstrative adjective
\item adj_ind: indefinite adjective
\item adj_int : interrogative adjective
\item adj_num: numerical adjective
\item adj_pos: possessive adjective
\item adj_sup:
}
\item adv: adverb
\itemize{
\item adv_sup
}
\item art_def: definite article
\item aux: auxiliary
\item con: conjunction
\item nom: nominal
\item ono: onomatopoeia
\item pre: preposition
\item pro: pronoun
\itemize{
\item pro_dem: demonstrative pronoun
\item pro_ind: indefinite pronoun
\item pro_per: personnal pronoun
\item pro_pos: possessive pronoun
\item pro_rel: relative pronoun
}
\item ver: verb
\itemize{
\item ver_sup:
}
}
}
\examples{
fr_expr=get_lexicon("fr")
}
