% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixr-package.R
\docType{package}
\name{mixr-package}
\alias{mixr}
\alias{mixr-package}
\title{mixr: Fit Mixed-Effects Models}
\description{
The mixr package fits linear mixed-effects models, centered around the 'nlme' package. It is centered around fitting many models of identical form, from a list of features as the response variable.
}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/0000-0002-1024-5859}{ORCID}) [copyright holder]

}
\keyword{internal}
\keyword{package}
