% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequencies.R
\name{plot_frequencies}
\alias{plot_frequencies}
\title{Returns a plot showing frequencies of cat (lemma or word for instance) in data}
\usage{
plot_frequencies(data, cat, frequency, scale = NA)
}
\arguments{
\item{data}{a tibble with frequencies of cat}

\item{cat}{words or lemmas, for instance}

\item{frequency}{frequency of cat}
}
\value{
a plot
}
\description{
Returns a plot showing frequencies of cat (lemma or word for instance) in data
}
\examples{
library(janeaustenr)
df <- tibble(txt=prideprejudice) \%>\% unnest_tokens(word,txt)
df_freq <- tidy_frequencies(df, word, criterion="n_min",n_min=500)
plot_frequencies(df_freq, word,n)
}
