% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_words.R
\name{plot_words}
\alias{plot_words}
\title{Plot a metric according to word (for instance, frequencies)}
\usage{
plot_words(df, words, metric, fill = NULL, fill_fixed = "grey50")
}
\arguments{
\item{df}{a tibble with words (or any string to be displayed) and metric (frequency for instance)}

\item{words}{the name of the column with words}

\item{metric}{the name of the column with the considered metric}
}
\value{
a plot
}
\description{
Plot a metric according to word (for instance, frequencies)
}
\examples{
mydf=dplyr::bind_rows(tibble::tibble(txt=janeaustenr::prideprejudice,
                                     novel="Pride and Prejudice"),
                      tibble::tibble(txt=janeaustenr::sensesensibility,
                                     novel="Sense and Sensibility")) \%>\%
  tidytext::unnest_tokens(word,txt)
df_freq <- tidy_frequencies(mydf,
                            cat=word,
                            top_freq=20)
plot_words(df_freq,word,freq)
}
