% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.selectEM.R
\name{print.selectEM}
\alias{print.selectEM}
\title{Print Method for Class \code{selectEM}}
\usage{
\method{print}{selectEM}(x, ...)
}
\arguments{
\item{x}{an object of class \code{selectEM}}

\item{...}{other arguments passed to \code{print}}
}
\description{
The function prints the result of mixture model selection.
}
\details{
The function \code{print.selectEM} is the print method for the class{selectEM}, which is the output
of the function \code{select}. It prints a data frame which contains the following information of each
candidate mixture models: the number of components, whether the variance is the same for each component
in a mixture model (only for normal), the value of BIC, and an indicator of the best model.
}
\seealso{
\code{\link{select}}
}
