% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateLongData.R
\name{simulateLongData}
\alias{simulateLongData}
\alias{createLongData}
\title{Simulate Longitudinal Data in R}
\usage{
simulateLongData(
  nsubj = 20,
  beta0 = 1000,
  beta1 = 10,
  max.obs = 10,
  group = NULL,
  auto.cor = 0.5,
  sd.pars = list(sigma = beta0/4, tau0 = 2, tau1 = 2, tau01 = 0.5),
  r.seed = 1234
)

createLongData(...)
}
\arguments{
\item{nsubj}{Integer. Number of subjects.}

\item{beta0}{Intercept coefficient.}

\item{beta1}{Slope coefficient.}

\item{max.obs}{Integer. Number of serial (longitudinal) observations per subject.}

\item{group}{Character. String to name the subject group type.}

\item{auto.cor}{The coefficient for the serial autocorrelation;
range \verb{[0, 1]}.}

\item{sd.pars}{A list containing:
\describe{
\item{sigma:}{true error SD.}
\item{tau0:}{intercept SD.}
\item{tau1:}{slope SD.}
\item{tau01:}{intercept-slope correlation.}
}}

\item{r.seed}{Numeric. Optional random number generator for
reproducibility of serial sample draws and RFU values.}

\item{...}{Arguments containing a list of parameters to be passed
to \code{\link[=simulateLongData]{simulateLongData()}}.}
}
\value{
A \code{tibble} of simulated data.

A \code{tibble} with a response and time/grouping fields.
}
\description{
This function creates simulated longitudinal data that is useful
for fitting of mixed-effects modeling, general longitudinal visualization,
and RFU analogues to serially dependent data. In general, this function
is intended to be used in support of \code{\link[=createLongData]{createLongData()}}, which
wraps this the simulation aspect into an overall data set
}
\section{Functions}{
\itemize{
\item \code{createLongData()}: To generate longitudinal data there is a wrapper around \code{\link[=simulateLongData]{simulateLongData()}}
that creates full data set with serial dependence for analysis, see examples.

}}
\examples{
long <- simulateLongData()  # default pars
dim(long)
head(long)
# single group: time response
p <- list(nsubj = 20, beta0 = 1000, beta1 = 400, max.obs = 10, r.seed = 101,
          sd.pars = list(sigma = 250), auto.cor = 0.1)
longData <- createLongData(subject = p)
head(longData)
table(longData$pid)

# simulate longitudinal data with group-specific response
p1 <- list(nsubj = 20, beta1 = 0, max.obs = 10, r.seed = 10,
           sd.pars = list(sigma = 150), auto.cor = 0.1)
p2 <- list(nsubj = 20, beta1 = 400, max.obs = 10, r.seed = 101,
           sd.pars = list(sigma = 350), auto.cor = 0.1)
groupResponseData <- createLongData(p1, p2)
head(groupResponseData)

# There are 2 ways to rename the group labels:
# rename via the '...'
groupResponseData <- createLongData(control = p1, treatment = p2)
head(groupResponseData)

# rename via parameters list
p1 <- c(p1, group = "control")
p2 <- c(p2, group = "treatiment")
groupResponseData <- createLongData(p1, p2)
head(groupResponseData)
}
\references{
\url{http://stats.stackexchange.com/questions/76999/simulating-longitudinal-lognormal-data-in-r}
}
\seealso{
\code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}
}
\author{
Stu Field
}
