% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Stamp.R
\docType{data}
\name{Stamp}
\alias{Stamp}
\title{1872 Hidalgo Stamp Data}
\format{
A vector with 485 measurements of the thickness (nm) of the stamps
}
\usage{
Stamp
}
\description{
A vector containing the 1872 Hidalgo stamp data
}
\references{
Izenman, A. J. and Sommer, C. J. Philatelic mixtures and multimodal densities.
\emph{Journal of the American Statistical association}, 83(404):941-953, 1988.
}
\keyword{datasets}
