% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Stamp2.R
\docType{data}
\name{Stamp2}
\alias{Stamp2}
\title{1872 Hidalgo Stamp Data (Binned)}
\format{
A matrix with 62 rows and 3 columns:
\describe{
 \item{lower}{the lower bin values}
 \item{upper}{the upper bin values}
 \item{freq}{the number of observations in each bin}
}
}
\usage{
Stamp2
}
\description{
A dataset containing the 1872 Hidalgo stamp data in the form of binned data
}
\keyword{datasets}
