% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_frequencies.R
\name{tidy_frequencies}
\alias{tidy_frequencies}
\title{Returns tibble with frequencies of cat (lemma or word for instance) in data}
\usage{
tidy_frequencies(data, cat, criterion = "top_n", top_n = 30, n_min = 10)
}
\arguments{
\item{data}{a tibble}

\item{cat}{words or lemmas, for instance}

\item{criterion}{should the output be filtered by top specificities by category ('top_n') or according to a minimum value of specificity ("spec_min").}

\item{top_n}{in case criterion=='top_n', how many items by category should be kept (defaults to 50)}

\item{spec_min}{in case criterion=='min_spec', which is the minimum specificity for an item to be kept (defaults to 2)}
}
\value{
a tibble of cat frequencies
}
\description{
Returns tibble with frequencies of cat (lemma or word for instance) in data
}
\examples{
library(janeaustenr)
df <- tibble(txt=prideprejudice) \%>\% unnest_tokens(word,txt)
tidy_frequencies(df, word, criterion="n_min",n_min=200)
}
