% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_geocode.R
\name{tidy_geocode}
\alias{tidy_geocode}
\title{Returns geographical information as a tibble through the use of a geocoding API}
\usage{
tidy_geocode(df, location, info = c("lat", "lng"), method = "opencage")
}
\arguments{
\item{df}{a tibble}

\item{location}{the variable to geocode}

\item{info}{the variables to inform: by default, latlng. Information can be e.g. "locality","country","country_code","formatted_address". See https://developers.google.com/maps/documentation/geocoding/intro#Types for the list of all possibilities supported by the Google Map geocoding API.}

\item{method}{the API to send the request to. Can be "opencage" (default) or "google".}
}
\value{
tibble with additional columns corresponding to the info required
}
\description{
Returns geographical information as a tibble through the use of a geocoding API
}
\examples{
mydf <-tibble(location=c("Lyon, France",
                         "22 place du Général de Gaulle, Paris",
                         "la Guillotière, Lyon",
                         NA,
                         "Europe",
                         "Tucson, AZ",
                         "Rio Grande do Sul",
                         ""))
tidy_geocode(mydf,location,method="opencage")
}
