% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_specificities.R
\name{tidy_specificities}
\alias{tidy_specificities}
\title{Returns a tibble with specificities according to two crossed categories.}
\usage{
tidy_specificities(mydf, cat1, cat2, top_spec = NA, min_spec = NA)
}
\arguments{
\item{mydf}{a tibble}

\item{cat1}{a factor corresponding to words or lemmas}

\item{cat2}{a category}

\item{top_spec}{how many items by category (filter based on specificity) should be kept. If not provided (the default) everything is kept.}

\item{min_spec}{which is the minimum specificity for an item to be kept. If not provided (the default) everything is kept.}
}
\value{
tibble with additional columns cat1, cat2, spec
}
\description{
Returns a tibble with specificities according to two crossed categories.
}
\examples{
 mydf=dplyr::bind_rows(
         tibble::tibble(txt=janeaustenr::prideprejudice,
         novel="Pride and Prejudice"),
         tibble::tibble(txt=janeaustenr::sensesensibility,
         novel="Sense and Sensibility")) \%>\%
      tidytext::unnest_tokens(word,txt)
 tidy_specificities(mydf,
                    cat1=word,
                    cat2=novel)
}
