% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lme-safely.R
\name{fit_lme_safely}
\alias{fit_lme_safely}
\title{Fit Linear Mixed-Effect Safely}
\usage{
fit_lme_safely(formula, ...)
}
\arguments{
\item{formula}{The model specification to fit in "formula" syntax,
according to \code{\link[=lme]{lme()}}.}

\item{...}{Additional arguments passed to \code{\link[=lme]{lme()}}.}
}
\value{
A linear mixed-effects model.
}
\description{
In multidimensional data analysis, fitting thousands of identical
mixed models often results in at least \emph{some} models failing
to converge on a final solution for the coefficients. This function
provides a convenient manner in which to bypass such cases and
continue fitting the remainder of the response covariates without
becoming hung up on convergence issues.
}
\details{
This function attempts to fit a linear mixed-effects model specified by
the argument \code{formula} via a specific sequence of methods:

\enumerate{
\item use the default options of \code{\link[nlme:lme]{nlme::lme()}}, if fails ...
\item use \code{method = "ML"} (Maximum Likelihood) option, if fails ...
\item use \code{method = "REML"}, restricted expectation maximum likelihood;
the default but with \code{control = lmeControl(opt = "optim")} option
(the old optimizer), if fails ...
\item use both \code{method = "ML"} and \code{control = lmeControl(opt = "optim")}
options, if fails ...
\item suppress warnings and simply return the \emph{non-converged}
object. Beware of this feature! Non-converged models should \emph{not}
be used in analyses. To identify these cases, a boolean entry named
"converged" is added to the model object indicating the convergence
status of the model.
}
}
\note{
The data used to fit the model is \emph{not} included
in the final model, as specified by the \code{keep.data = FALSE}
argument in \code{\link[=lme]{lme()}}.
}
\examples{
long <- simulateLongData(r.seed = 101)
fit  <- fit_lme_safely(yij ~ time, random = ~1|pid, data = long)
summary(fit)
anova(fit)
}
\seealso{
\code{\link[=lme]{lme()}}
}
\author{
Stu Field
}
