% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedEffectsModels.R,
%   R/createMixedEffectsTable.R
\name{fitMixedEffectsModels}
\alias{fitMixedEffectsModels}
\alias{createMixedEffectsTable}
\alias{print.mixed_effects_table}
\title{Fit Mixed-Effects Models}
\usage{
fitMixedEffectsModels(
  data,
  fixed = "TimePoint*SampleGroup",
  random = "~ 1 | pid",
  do.log
)

createMixedEffectsTable(x)

\method{print}{mixed_effects_table}(x, n = 6, ...)
}
\arguments{
\item{data}{The \code{data.frame} containing RFU data to as columns.}

\item{fixed}{Character. Indicating the right-hand-side of the
formula for the fixed effect.}

\item{random}{Character. Indicating the full formula for the random
effect. The default fits subject specific intercepts. Typically "group"
field indicator (e.g. subjects) are to the right side of the "|" and this
field \emph{must} be a column name in \code{data}.}

\item{do.log}{Deprecated. Please log-transform prior to fitting.}

\item{x}{A list of mixed-effects models calculated
using \code{\link[=fitMixedEffectsModels]{fitMixedEffectsModels()}}. Alternatively, an object of
class \code{"mixed_effects_table"} for the S3 print method.}

\item{n}{Numeric. The number of rows to show in the S3 print method.}

\item{...}{Additional arguments as required by default S3 print methods.}
}
\value{
A list (invisibly) of fitted linear mixed-effects response models
(\code{class = lme}) with extra attributes attached for print methods.

A list object of class \code{mixed_effects_table} with the
following components:
\item{stat.table}{A data frame (table) of the results:
\describe{
\item{Time effect:}{Both F-values and p-values for TimePoint}
\item{Group effect:}{Both F-values and p-values for Group}
\item{p.value:}{Unadjusted p-values for the KS-distances}
\item{fdr:}{FDR adjustment to p-values}
\item{p.bonferroni:}{Bonferroni corrected p-value}
\item{rank:}{Ranking by the test significance.}
}}
\item{test}{The name of the model fit, "Linear Mixed-effects Model".}
\item{call}{The direct call made the \code{\link[=fitMixedEffectsModels]{fitMixedEffectsModels()}}.}
\item{data.dim}{The dimensions of the data frame.}
\item{method}{The fitting method used by \code{lme}.}
\item{fixed}{The fixed-effects.}
\item{observations}{The number of total observations.}
\item{PIDfield}{The field name identifying the subjects in the data frame.}
\item{subjects}{The number of subjects.}
}
\description{
Fit an identical series of mixed-effects models according
for each analyte in an ADAT.

The ordering of the table is determined by the fixed effects
argument passed to \code{\link[=fitMixedEffectsModels]{fitMixedEffectsModels()}}. Generally,
if there is an interaction term, that p-value is used, otherwise
the other fixed effects, usually \code{TimePoint}.
}
\details{
By default, random intercept models (subject specific offsets)
are fit with a fixed effect for slope: \code{random = ~ 1 | group}.
This means that group/subject specific slopes are not fit, and all
fits will be parallel with respect to slope. This may or may not
be what you intend but for most of SomaScan data this is the
desired default.
}
\section{Functions}{
\itemize{
\item \code{print(mixed_effects_table)}: There is a S3 print method for \code{mixed_effects_table} class.

}}
\examples{
df <- data.frame(
  pid       = 1041:1080,
  Pop       = rep(utils::head(LETTERS, 10), 4),
  TimePoint = rep(c("baseline", "6 months", "12 months", "24 months"), each = 10),
  seq.1234.56 = stats::rnorm(40, mean = 25, sd = 3.5),
  seq.6969.4  = stats::rnorm(40, mean = 25, sd = 1.5),
  Response  = factor(sample(c("Control", "Disease"), 40, replace = TRUE))
)
df$TimePoint <- factor(df$TimePoint, levels = c("baseline", "6 months",
                                                "12 months", "24 months"))

models <- df[, c("TimePoint", "Response", "Pop", mixr:::get_analytes(df))] |>
  fitMixedEffectsModels(fixed = "TimePoint*Response", random = "~ 1 | Pop")
lapply(models, summary)

# `soma_lme` class object
class(models)

# create table from `soma_lme` object
lme_tab <- createMixedEffectsTable(models)
# S3 print method
lme_tab

}
\seealso{
\code{\link[=fit_lme_safely]{fit_lme_safely()}}
}
\author{
Stu Field
}
