% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme-diagnostic.R
\name{lme_diagnostic}
\alias{lme_diagnostic}
\title{Diagnostics of Linear Mixed-Effects Models}
\usage{
lme_diagnostic(model, data = NULL, group_by = NULL, ...)
}
\arguments{
\item{model}{A \code{lme} model object, typically fit via \code{\link[=fit_lme_safely]{fit_lme_safely()}}.}

\item{data}{Optionally, if the data used in the lme fit is not
included with the model, this must be passed here.}

\item{group_by}{Optional. If passed, either a vector indicating the
split in the subjects (i.e. groups) or a character field string
indicating the column in \code{data} containing the split information.}

\item{...}{Additional arguments passed to \code{\link[nlme:lmList]{nlme::lmList()}}.}
}
\value{
Diagnostic plots (see description) showing the subject specific
linear model coefficients (slope and intercept).
}
\description{
A diagnostic wrapper for linear mixed-effects models. Generally for
individual models fit at the command line, it produces the individual
subject-specific (group-specific) slopes and intercepts (offset) for serially
dependent (longitudinal) data. Two independent plots are generated:
\describe{
\item{Intervals}{A plot of the 95\% confidence interval
for each parameter estimate by subject.}
\item{Boxplots}{Boxplots of the estimated offsets (bottom) and
slopes (top). Optionally, the boxplots cad be split by a grouping
variable either passed as a vector, or included as part of the data.}
}
}
\examples{
p1 <- list(beta1 = 50, r_seed = 101)
p2 <- list(beta1 = 10, r_seed = 405)
long_data <- create_long_data(A = p1, B = p2)
fit <- fit_lme_safely(yij ~ time, random = ~1|pid, data = long_data)
lme_diagnostic(fit, long_data)  # all groups/pids together

# Group boxplots by sub-group
lme_diagnostic(fit, long_data, group_by = "Group")  # slope A > B

# Group by external random vector (3 levels)
group_vec <- withr::with_seed(5, sample(1:3, nrow(long_data), replace = TRUE))
lme_diagnostic(fit, long_data, group_by = group_vec)
}
\seealso{
\code{\link[nlme:lmList]{nlme::lmList()}}, \code{\link[nlme:intervals]{nlme::intervals()}}, \code{\link[=fit_lme_safely]{fit_lme_safely()}}
}
