% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixr.R
\docType{package}
\name{mixr}
\alias{mixr}
\title{\code{mixr} package}
\description{
A package that mixes a few tools to analyse texts
}
\details{
See the README on
\href{https://cran.r-project.org/package=mixr/README.html}{CRAN}
or \href{https://github.com/lvaudor/mixr#readme}{GitHub}
}
