% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixr.R
\docType{package}
\name{mixr}
\alias{mixr}
\title{\code{mixr} package}
\details{
Functions to simplify text analysis with multiple languages.
Uses Lexique382 dictionnaries (French lexicon and expressions) and Iramuteq's dictionnaries for other languages.
See the README on
\href{https://github.com/lvaudor/mixr#readme}{GitHub}
}
