% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_specificities.R
\name{plot_specificities}
\alias{plot_specificities}
\title{Plot specificities of cat1 (lemma or word) according to cat2 (categories)}
\usage{
plot_specificities(spec_data, cat1, cat2, criterion = "top_n", top_n = 50,
  min_spec = 2)
}
\arguments{
\item{cat1}{words or lemmas}

\item{cat2}{categories}

\item{criterion}{should the information displayed be filtered by top specificities by category ('top_n') or according to a minimum value of specificity ("spec_min").}

\item{top_n}{in case criterion=='top_n', how many items by category should be kept (defaults to 50)}

\item{data}{a tibble}

\item{spec_min}{in case criterion=='min_spec', which is the minimum specificity for an item to be kept (defaults to 2)}
}
\value{
a plot
}
\description{
Plot specificities of cat1 (lemma or word) according to cat2 (categories)
}
\examples{
"pouet"
}
