% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_specificities.R
\name{plot_specificities}
\alias{plot_specificities}
\title{Plot specificities of cat1 (lemma or word) according to cat2 (categories)}
\usage{
plot_specificities(df_spec, cat1, cat2)
}
\arguments{
\item{df_spec}{a tibble with specifictities of cat1 according to cat2}

\item{cat1}{words or lemmas}

\item{cat2}{categories}
}
\value{
a plot
}
\description{
Plot specificities of cat1 (lemma or word) according to cat2 (categories)
}
\examples{
mydf=dplyr::bind_rows(tibble::tibble(txt=janeaustenr::prideprejudice,
                                     novel="Pride and Prejudice"),
                      tibble::tibble(txt=janeaustenr::sensesensibility,
                                     novel="Sense and Sensibility")) \%>\%
 tidytext::unnest_tokens(word,txt)
df_spec <- tidy_specificities(mydf,
                              cat1=word,
                              cat2=novel,
                              min_spec=5)
plot_specificities(df_spec,
                   cat1=word,
                   cat2=novel)
}
