% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_specificities.R
\name{plot_specificities}
\alias{plot_specificities}
\title{Plot specificities of cat1 (lemma or word) according to cat2 (categories)}
\usage{
plot_specificities(spec_data, cat1, cat2)
}
\arguments{
\item{cat1}{words or lemmas}

\item{cat2}{categories}

\item{data}{a tibble}
}
\value{
a plot
}
\description{
Plot specificities of cat1 (lemma or word) according to cat2 (categories)
}
\examples{
"pouet"
}
