% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-lme.R
\name{simple_lme}
\alias{simple_lme}
\alias{plot.simple_LME}
\alias{print.simple_LME}
\title{Create Simple lme Models}
\usage{
simple_lme(data, response, fixed, grouping, random_slope = TRUE, ...)

\method{plot}{simple_LME}(x, col = "blue", pt_col = "black", groups = NULL, ...)

\method{print}{simple_LME}(x, ...)
}
\arguments{
\item{data}{Either a data frame or a \emph{named} list of
\code{data.frame} objects. If a data frame is provided then a single lme model is
built for each response in \code{responses}. If a named list is provided
then a distinct model is built for each \verb{data.frame x response} combination.}

\item{response}{Character. A vector with the names of the response
variables.}

\item{fixed}{Character. The name of the fixed covariate variable
(e.g. \code{TimePoint}), typically the slope component of the model.}

\item{grouping}{Character. The name of the variable indicating the groups
(e.g. \code{pid} or \code{subjectID}).}

\item{random_slope}{Logical. Indicating whether subject specific slopes
should be fit. Alternatively, subject specific offsets (intercept) are fit.}

\item{...}{Additional arguments passed to \code{\link[=fit_lme_safely]{fit_lme_safely()}}
or is via the S3 plot method, passed to
\code{graphics::\link[graphics]{plot}}.}

\item{x}{An \code{simple_LME} class object.}

\item{col}{The color of the lines. See \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}.}

\item{pt_col}{The color of the points. See \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{groups}{\code{numeric(n)}. Indices indicating the \emph{optional} subset of
groups (e.g. subjects) to be plotted for the random effects.}
}
\value{
A list with an element for each response in \code{responses} as well
as the fixed and grouping variables. If \code{data} is a data.frame or a
groupedData object then each of the response elements will be an lme model.
Otherwise if a list was provided for the \code{data} argument then the
response elements will be a list with an lme model for each data set in the
\code{data} list.
}
\description{
Builds a series linear mixed-effects models of the form \verb{Response ~ Fixed Covariate | Grouping}. A single model or a set of models is built
for each response in \code{responses}. The input data is stored along
with names of the fixed covariate and grouping variables.
}
\section{Functions}{
\itemize{
\item \code{plot(simple_LME)}: S3 method plots the points and fit for each of the
random effects in a \code{\link[=simple_lme]{simple_lme()}} model object.

\item \code{print(simple_LME)}: S3 method print method for \code{simple_LME} model object.

}}
\examples{
long  <- simulate_long_data(r_seed = 101, beta1 = 50)
model <- simple_lme(long, "yij", "time", "pid")
model   # S3 print method

# random offset
model2 <- simple_lme(long, "yij", "time", "pid", random_slope = FALSE)

# 2 data set option
long2  <- simulate_long_data(r_seed = 405, beta1 = 10)     # smaller beta1
model3 <- simple_lme(list(A = long, B = long2), "yij", "time", "pid")

# S3 print method
model3

# S3 plot method
plot(model)

# plot only certain subjects/groups
plot(model, groups = seq(1, 11, by = 2), pt.col = "purple")
}
\seealso{
\code{\link[=fit_lme_safely]{fit_lme_safely()}}
}
\author{
Michael R. Mehan

Michael R. Mehan, Stu Field
}
