% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-long-data.R
\name{simulate_long_data}
\alias{simulate_long_data}
\alias{create_long_data}
\title{Simulate Longitudinal Data in R}
\usage{
simulate_long_data(
  nsubj = 20L,
  beta0 = 1000,
  beta1 = 10,
  max_obs = 10,
  group = NULL,
  auto_cor = 0.5,
  sd_pars = list(sigma = beta0/4, tau0 = 2, tau1 = 2, tau01 = 0.5),
  r_seed = 1234L
)

create_long_data(...)
}
\arguments{
\item{nsubj}{\code{integer(1)}. Number of subjects.}

\item{beta0}{\code{numeric(1)}. The intercept coefficient.}

\item{beta1}{\code{numeric(1)}. The slope coefficient.}

\item{max_obs}{\code{integer(1)}. Number of serial (longitudinal)
observations per subject.}

\item{group}{\code{character(1)}. String to name the subject group type.}

\item{auto_cor}{\code{numeric(1)}. The coefficient for the serial
autocorrelation in \verb{[0, 1]}.}

\item{sd_pars}{A list (all \code{numeric(1)}) containing:
\describe{
\item{sigma:}{true error SD.}
\item{tau0:}{intercept SD.}
\item{tau1:}{slope SD.}
\item{tau01:}{intercept-slope correlation.}
}}

\item{r_seed}{\code{integer(1)}. Random number generator for
reproducibility of serial sample draws.}

\item{...}{Arguments containing a list of parameters to be passed
to \code{\link[=simulate_long_data]{simulate_long_data()}}.}
}
\value{
A \code{tibble} of simulated data.

A \code{tibble} with a response and time/grouping fields.
}
\description{
This function creates simulated longitudinal data that is useful
for fitting of mixed-effects modeling, general longitudinal
visualization, and feature analogues to serially dependent data.
In general, this function is intended to be used in support
of \code{\link[=create_long_data]{create_long_data()}}, which wraps this the simulation aspect
into an overall data set.
}
\section{Functions}{
\itemize{
\item \code{create_long_data()}: To generate longitudinal data there is a wrapper around \code{\link[=simulate_long_data]{simulate_long_data()}}
that creates full data set with serial dependence for analysis, see examples.

}}
\examples{
long <- simulate_long_data()  # default pars
dim(long)
head(long)
# single group: time response
p <- list(nsubj = 20, beta0 = 1000, beta1 = 400, max_obs = 10, r_seed = 101,
          sd_pars = list(sigma = 250), auto_cor = 0.1)
long_data <- create_long_data(subject = p)
head(long_data)
table(long_data$pid)

# simulate longitudinal data with group-specific response
p1 <- list(nsubj = 20, beta1 = 0, max_obs = 10, r_seed = 10,
           sd_pars = list(sigma = 150), auto_cor = 0.1)
p2 <- list(nsubj = 20, beta1 = 400, max_obs = 10, r_seed = 101,
           sd_pars = list(sigma = 350), auto_cor = 0.1)
groupResponseData <- create_long_data(p1, p2)
head(groupResponseData)

# There are 2 ways to rename the group labels:
# rename via the '...'
groupResponseData <- create_long_data(control = p1, treatment = p2)
head(groupResponseData)

# rename via parameters list
p1 <- c(p1, group = "control")
p2 <- c(p2, group = "treatiment")
groupResponseData <- create_long_data(p1, p2)
head(groupResponseData)
}
\references{
\url{http://stats.stackexchange.com/questions/76999/simulating-longitudinal-lognormal-data-in-r}
}
\seealso{
\code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}
}
\author{
Stu Field
}
