% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_frequencies.R
\name{tidy_frequencies}
\alias{tidy_frequencies}
\title{Returns tibble with frequencies of cat (lemma or word for instance) in data}
\usage{
tidy_frequencies(mydf, cat, top_freq = NA, min_freq = NA)
}
\arguments{
\item{mydf}{a tibble}

\item{cat}{words or lemmas, for instance}

\item{top_freq}{how many items by category (filter based on frequency) should be kept. If not provided (the default) everything is kept.}

\item{min_freq}{which is the minimum specificity for an item to be kept. If not provided (the default) everything is kept.}
}
\value{
a tibble of cat frequencies
}
\description{
Returns tibble with frequencies of cat (lemma or word for instance) in data
}
\examples{
mydf <- tibble::tibble(txt=janeaustenr::prideprejudice) \%>\%
  tidytext::unnest_tokens(word,txt)
tidy_frequencies(mydf, word, min_freq=200)
}
