% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_geocode.R
\name{tidy_geocode}
\alias{tidy_geocode}
\title{Returns geographical information as a tibble through the use of a geocoding API}
\usage{
tidy_geocode(df, location, info = c("lat", "lng"), method = "opencage")
}
\arguments{
\item{df}{a tibble}

\item{location}{the variable to geocode}

\item{info}{the variables to inform
Information can be many other variables depending on the method you are using.
Type \code{help(geocode_google)} or \code{help(geocode_opencage)} (according to the method you wish to use) to see details.}

\item{method}{the API to send the request to. Can be "opencage" (default) or "google".}
}
\value{
tibble with additional columns corresponding to the info required
}
\description{
Returns geographical information as a tibble through the use of a geocoding API
}
\examples{
mydf <-tibble::tibble(location=c("Lyon, France",
                                 "22 place du Général de Gaulle, Paris",
                                 "la Guillotière, Lyon",
                                 NA,
                                 "Europe",
                                 "Tucson, AZ",
                                 "Rio Grande do Sul",
                                 ""))
tidy_geocode(mydf,location,method="opencage")
}
