% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_geocode.R
\name{tidy_geocode}
\alias{tidy_geocode}
\title{Returns a tibble that includes a variable describing locations, with added information corresponding to the Data Science Toolkit geocoding tool.}
\usage{
tidy_geocode(data, location)
}
\arguments{
\item{data}{a tibble}

\item{location}{the variable to geocode}
}
\value{
tibble with additional columns lat, lng, country, locality, area
}
\description{
Returns a tibble that includes a variable describing locations, with added information corresponding to the Data Science Toolkit geocoding tool.
}
\examples{
df <-tibble(loc=c("Lyon, France",
                  "22 place du Général de Gaulle, Paris",
                  "la Guillotière, Lyon",
                  "Europe",
                  "Tucson, AZ",
                  "Rio Grande do Sul"))
tidy_geocode(df,loc)
}
