% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_specificities.R
\name{tidy_specificities}
\alias{tidy_specificities}
\title{Returns a tibble with specificities according to two crossed categories.}
\usage{
tidy_specificities(data, cat1, cat2)
}
\arguments{
\item{data}{a tibble}

\item{cat1}{a category}

\item{cat2}{a factor describing corresponding to words or lemmas}
}
\value{
tibble with additional columns cat1, cat2, spec
}
\description{
Returns a tibble with specificities according to two crossed categories.
}
\examples{
"pouet"
}
